Fix : table route
